+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.5.0
+
+* Added `Jump.CredoChecks.NoManualContentDisposition` check, courtesy of @ftes ([PR](https://github.com/Jump-App/credo_checks/pull/25).
+* Improvement to `Jump.CredoChecks.UnusedLiveViewAssign`: Add `custom_assign_functions` to track project-specific helpers that write assigns (`{function, arity}` or `{module, function, arity}`).
+
 ## v0.4.0
 
 Added 4 new checks: